requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
django-extensions==3.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
mailchimp-marketing==3.0.80
python-decouple==3.8
Pillow==10.1.0
reportlab==4.0.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-filter==23.5
django-tables2==2.7.0
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
openpyxl==3.1.2
django-import-export==3.3.4
django-guardian==2.4.0
whitenoise==6.6.0
gunicorn==21.2.0